+++ app/App.js
@@ -3,12 +3,14 @@ import { Switch, Route } from 'react-router-dom'
 
 import PrivateRoute from 'components/PrivateRoute'
 import NotFound from 'components/NotFound'
+import Layout from 'components/Layout'
+import HomePage from 'components/HomePage'
 
 import routes from './routes'
-import Layout from './components/Layout'
 
 const App = () => (
   <Switch>
+    <Route path="/" exact component={HomePage} />
     {routes.map(route =>
       route.private ? (
         <PrivateRoute
