+++ app/components/HomePage/index.js
@@ -4,12 +4,15 @@ import Hero from '../Hero/Hero'
 import PageSection from '../PageSection/PageSection'
 import ContentContainer from '../ContentContainer/ContentContainer'
 import ActivityList from '../Activity/ActivityList'
+import Footer from 'components/Footer'
+import Header from 'components/Header'
 
-import './HomePage.css'
+import './style.css'
 
 const HomePage = () => {
   return (
     <Fragment>
+      <Header theme="inverse" />
       <Hero />
       <PageSection>
         <ContentContainer>
@@ -34,6 +37,7 @@ const HomePage = () => {
           <p>PageSection</p>
         </ContentContainer>
       </PageSection>
+      <Footer theme="inverse" />
     </Fragment>
   )
 }
