+++ app/components/PartnersPage/PartnersPage.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import Header from '../Header/Header'
-import Partners from '../Partners/Partners'
+import Header from 'components/Header'
+import Partners from 'components/Partners/Partners'
 
 const PartnersPage = () => {
   return (
